generic search

+++ funtime/search-models/src/main/java/com/example/models/City.java
@@ -16,5 +16,5 @@ public class City
     private String favCity;
     private Boolean active;
     private String state;
-    private Integer pin;
+    private Long pin;
 }
